delete unnecesary code
