+++ core/webcam.py
@@ -1,0 +1,85 @@
+import asyncio
+import json
+import os
+
+from aiohttp import web
+from aiortc import RTCPeerConnection, RTCSessionDescription, RTCConfiguration
+from aiortc.contrib.media import MediaPlayer, MediaRelay
+
+ROOT = os.path.dirname(__file__)
+HOST = "0.0.0.0"
+PORT = 8080
+
+
+relay = None
+webcam = None
+
+
+def create_local_tracks():
+    global relay, webcam
+    if relay is None:
+        options = {"framerate": "15", "video_size": "640x480"}
+        webcam = MediaPlayer("/dev/video0", format="v4l2", options=options)
+        relay = MediaRelay()
+    return relay.subscribe(webcam.video)
+
+
+
+async def index(request):
+    content = open(os.path.join(ROOT, "index.html"), "r").read()
+    return web.Response(content_type="text/html", text=content)
+
+
+async def javascript(request):
+    content = open(os.path.join(ROOT, "client.js"), "r").read()
+    return web.Response(content_type="application/javascript", text=content)
+
+
+async def offer(request):
+    params = await request.json()
+    offer = RTCSessionDescription(sdp=params["sdp"], type=params["type"])
+
+    pc = RTCPeerConnection(configuration=RTCConfiguration(iceServers=[]))
+    pcs.add(pc)
+
+    @pc.on("connectionstatechange")
+    async def on_connectionstatechange():
+        print("Connection state is %s" % pc.connectionState)
+        if pc.connectionState == "failed":
+            await pc.close()
+            pcs.discard(pc)
+
+    video = create_local_tracks()
+
+    pc.addTrack(video)
+
+    await pc.setRemoteDescription(offer)
+
+    answer = await pc.createAnswer()
+    await pc.setLocalDescription(answer)
+
+    return web.Response(
+        content_type="application/json",
+        text=json.dumps(
+            {"sdp": pc.localDescription.sdp, "type": pc.localDescription.type}
+        ),
+    )
+
+
+pcs = set()
+
+
+async def on_shutdown(app):
+    # close peer connections
+    coros = [pc.close() for pc in pcs]
+    await asyncio.gather(*coros)
+    pcs.clear()
+
+
+if __name__ == "__main__":
+    app = web.Application()
+    app.on_shutdown.append(on_shutdown)
+    app.router.add_get("/", index)
+    app.router.add_get("/client.js", javascript)
+    app.router.add_post("/offer", offer)
+    web.run_app(app, host=HOST, port=PORT)
